split video player into seperate file with use client
